+++ pkg/controller/util.go
@@ -111,7 +111,7 @@ func removeAddressFromLoadBalancerIngress(a []v1.LoadBalancerIngress, addr strin
 // targetPort is a number, use that.  If the targetPort is a string, look that
 // string up in all named ports in all containers in the target pod.  If no
 // match is found, fail.
-func podFindPort(pod *v1.Pod, svcPort *v1.ServicePort) (int, error) {
+func podFindPort(pod *v1.Pod, svcPort *v1.ServicePort) (int32, error) {
 	portName := svcPort.TargetPort
 	switch portName.Type {
 	case intstr.String:
@@ -119,12 +119,12 @@ func podFindPort(pod *v1.Pod, svcPort *v1.ServicePort) (int, error) {
 		for _, container := range pod.Spec.Containers {
 			for _, port := range container.Ports {
 				if port.Name == name && port.Protocol == svcPort.Protocol {
-					return int(port.ContainerPort), nil
+					return port.ContainerPort, nil
 				}
 			}
 		}
 	case intstr.Int:
-		return portName.IntValue(), nil
+		return int32(portName.IntValue()), nil
 	}
 
 	return 0, fmt.Errorf("no suitable port for manifest: %s", pod.UID)
